필수 요구사항 완료

+++ src/main/java/com/sparta/calender/service/CalenderService.java
@@ -37,28 +37,40 @@ public class CalenderService {
     }
 
     @Transactional
-    public Long updateCalender(Long id, CalenderRequestDto calenderRequestDto) {
+    public boolean updateCalender(Long id, CalenderRequestDto calenderRequestDto) {
         // 해당 메모가 DB에 존재하는지 확인
         Calender calender = findCalender(id);
 
-        calender.update(calenderRequestDto);
-        return id;
-
+        if(calenderRequestDto.getPassword().equals(calender.getPassword())){
+            calender.update(calenderRequestDto);
+            return true;
+        }else{
+            return false;
+        }
     }
 
-    public Long deleteCalender(Long id) {
+    public boolean deleteCalender(Long id, String password) {
         // 해당 메모가 DB에 존재하는지 확인
         Calender calender = findCalender(id);
 
-        calenderRepository.delete(calender);
-        return id;
-
+        if(password.equals(calender.getPassword())){
+            calenderRepository.delete(calender);
+            return true;
+        }else{
+            return false;
+        }
     }
 
     private Calender findCalender(Long id) {
+        System.out.println("findCalender 메소드 진입 성공, id : "+id);
         return calenderRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("선택한 일정은 존재하지 않습니다.")
         );
     }
 
+    public CalenderResponseDto showCalenderDetails(Long id) {
+        System.out.println("service 진입 성공, id : "+id);
+        Calender calender = findCalender(id);
+        return new CalenderResponseDto(calender);
+    }
 }
